pr for spring upgrade (#40)
* changes made for czentrix

* spring version upgrade and compilation error resolved

+++ src/main/java/com/iemr/inventory/controller/stockExit/StockExitController.java
@@ -192,7 +192,7 @@ public class StockExitController {
 
 		try {
 
-			T_PatientIssue getData = stockExitService.getPatientissueAllDetail(itemStockinput.getPatientIssueID());
+			T_PatientIssue getData = stockExitService.getPatientissueAllDetail((Long) itemStockinput.getPatientIssueID());
 
 			response.setResponse(getData.toString());
